+++ HISTORY.rst
@@ -8,6 +8,7 @@ UNRELEASED
 
 * Fix extension loading for python3
 * Add a make target for building distributions (sdist + wheels)
+* Fix eval conversion for python 3
 
 0.1.2 (2016-08-03)
 ---------------------